+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add `akash_providers` datasource
+### Changed
+- `akash_deployment` output `services` now has a new structure and more information regarding replicas
+### Development
+- Integration with Praetor's API
+
 ## [0.0.6] - 2022-11-16
 ### Added
 - Add support to mainnet-4